+++ src/org/andreschnabel/jprojectinspector/metrics/project/Issues.java
@@ -51,7 +51,7 @@ public class Issues implements OnlineMetric {
 	}
 
 	@Override
-	public float measure(Project p) throws Exception {
+	public double measure(Project p) throws Exception {
 		return getNumberOfIssues(p);
 	}
 }
