+++ src/org/andreschnabel/jprojectinspector/metrics/project/ProjectAge.java
@@ -38,7 +38,7 @@ public class ProjectAge implements OnlineMetric {
 	}
 
 	@Override
-	public float measure(Project p) throws Exception {
+	public double measure(Project p) throws Exception {
 		return getProjectAge(p);
 	}
 }
