+++ src/org/andreschnabel/jprojectinspector/metrics/project/RecentCommits.java
@@ -44,7 +44,7 @@ public class RecentCommits implements OnlineMetric {
 	}
 
 	@Override
-	public float measure(Project p) throws Exception {
+	public double measure(Project p) throws Exception {
 		return getNumOfRecentCommits(p);
 	}
 }
